requirements.txt
# Medical Cancer CNN — Python 3.9+
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.5.0
tqdm>=4.65.0
# For download_pcam.py (PatchCamelyon real dataset)
h5py>=3.8.0
gdown>=4.6.0
